Fixed serialization with a bunch of things
Added the ability to save transform references, this is so that
when a script holds a reference to something else in the scene, it can
be saved though the transform reference. Added it so that the AIs are
saved and loaded correctly. Also changed indenting to tabs with some
of the files.