Fixed interacting with objects and added note close
The interacting with objects I fixed is for some reason the way
that I was getting a ray for the camera and the cursor gave be
a direction that was off by 90∘ so I changed to another way.
This was a pain to fix.

I also added a button to close the
currently open note, this is important because if you close with
escape instead then it opens the menu, which if you click on the
resume button then it clicks the note, which repeats, which is
not good.